requirements.txt
elevenlabs==1.8.1
openai>=1.0.0
python-dotenv>=1.0.0
pathlib>=1.0.0
dataclasses>=0.6
flask>=2.0.0
flask-cors>=3.0.0